You can see the traces using gtwkave
I had some problems. First up the avr sim only seems to want to IO ports
pins. It'd be cool if it was all memory as I'd assumed :D

Also I learnt that every IO port that can be written / read to with OUT
and IN has two addreses

- An in memory address which is IO Port Addres + 0x20
- And a port address

AvrSim assumes you're referencing using the memory address and subs
0x20 from it before adding it as a trace. It doesn't check to see if the
address you're adding is < 0x20. If it is you get a nasty seg fault.

I fixed this on my version so it doesn't add the trace and warns